# Conflicts:
#	ruoyi-admin/src/main/java/com/ruoyi/web/controller/common/CaptchaController.java
#	ruoyi-common/src/main/java/com/ruoyi/common/constant/Constants.java
#	ruoyi-common/src/main/java/com/ruoyi/common/core/domain/entity/SysUser.java
#	ruoyi-common/src/main/java/com/ruoyi/common/extension/MpSqlInjector.java
#	ruoyi-common/src/main/java/com/ruoyi/common/utils/file/FileUtils.java
#	ruoyi-framework/src/main/java/com/ruoyi/framework/config/MybatisPlusConfig.java
#	ruoyi-generator/src/main/java/com/ruoyi/generator/util/VelocityInitializer.java
#	ruoyi-generator/src/main/resources/vm/java/mapper.java.vm
#	ruoyi-generator/src/main/resources/vm/java/serviceImpl.java.vm
#	ruoyi-generator/src/main/resources/vm/xml/mapper.xml.vm

+++ ruoyi-system/src/main/java/com/ruoyi/system/service/impl/SysPostServiceImpl.java
@@ -67,7 +67,7 @@ public class SysPostServiceImpl implements ISysPostService
      * @return 选中岗位ID列表
      */
     @Override
-    public List<Integer> selectPostListByUserId(Long userId)
+    public List<Long> selectPostListByUserId(Long userId)
     {
         return postMapper.selectPostListByUserId(userId);
     }
@@ -137,7 +137,6 @@ public class SysPostServiceImpl implements ISysPostService
      * 
      * @param postIds 需要删除的岗位ID
      * @return 结果
-     * @throws Exception 异常
      */
     @Override
     public int deletePostByIds(Long[] postIds)
